#12 Create a ThresholdFinder base trait with common code. Moved log2 and entropy to ThresholdFinder because that is the only place that uses them.
