+++ src/main/scala/org/apache/spark/mllib/feature/FeatureUtils.scala
@@ -28,24 +28,6 @@ import scala.collection.mutable
 @Experimental
 object FeatureUtils {
 
-
-  private val LOG2 = math.log(2)
-
-  /** @return log base 2 of x */
-  val log2 = { x: Double => math.log(x) / LOG2 }
-
-  /**
-    * @param freqs sequence of integer frequencies.
-    * @param n the sum of all the frequencies in the list.
-    * @return the total entropy
-    */
-  def entropy(freqs: Seq[Long], n: Long) = {
-    -freqs.aggregate(0.0)(
-      { case (h, q) => h + (if (q == 0) 0 else (q.toDouble / n) * log2(q.toDouble / n))},
-      { case (h1, h2) => h1 + h2 }
-    )
-  }
-
   /**
    * Returns a vector with features filtered.
    * Preserves the order of filtered features the same as their indices are stored.
